# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EduBridge Egypt — Education platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /courses/{courseId}/assignments:
    get:
      operationId: listAssignments
      tags: [assignments]
      summary: List assignments in a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assignment"
    post:
      operationId: createAssignment
      tags: [assignments]
      summary: Create assignment
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssignmentBody"
      responses:
        "201":
          description: Created assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
  /assignments/{assignmentId}:
    get:
      operationId: getAssignment
      tags: [assignments]
      summary: Get assignment
      parameters:
        - name: assignmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
    patch:
      operationId: updateAssignment
      tags: [assignments]
      summary: Update assignment
      parameters:
        - name: assignmentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAssignmentBody"
      responses:
        "200":
          description: Updated assignment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"

  # ── SUBMISSIONS ───────────────────────────────────────────────────────────
  /assignments/{assignmentId}/submissions:
    get:
      operationId: listSubmissions
      tags: [submissions]
      summary: List submissions for assignment
      parameters:
        - name: assignmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Submission"
    post:
      operationId: submitAssignment
      tags: [submissions]
      summary: Submit assignment
      parameters:
        - name: assignmentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubmissionBody"
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
  /submissions/{submissionId}/grade:
    patch:
      operationId: gradeSubmission
      tags: [submissions]
      summary: Grade a submission
      parameters:
        - name: submissionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GradeSubmissionBody"
      responses:
        "200":
          description: Graded submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"

  # ── ANNOUNCEMENTS ─────────────────────────────────────────────────────────
...
  /analytics/dashboard:
    get:
      operationId: getDashboardSummary
      tags: [analytics]
      summary: Dashboard summary — totals, counts, trends
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
        - name: role
          in: query
          required: true
          schema:
            type: string
            enum: [teacher, student, parent, admin]
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /analytics/student/{studentId}/progress:
    get:
      operationId: getStudentProgress
      tags: [analytics]
      summary: Student learning progress per course
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Progress
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentProgress"
  /analytics/teacher/{teacherId}/revenue:
    get:
      operationId: getTeacherRevenue
      tags: [analytics]
      summary: Teacher monthly revenue breakdown
      parameters:
        - name: teacherId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Revenue data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RevenuePoint"
...
  /analytics/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [analytics]
      summary: Recent platform activity feed
      parameters:
        - name: userId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Activity feed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
  /parents/{parentId}/children:
    get:
      operationId: getParentChildren
      tags: [parents]
      summary: Get children linked to parent
      parameters:
        - name: parentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Children list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /parents/{parentId}/link-child:
    post:
      operationId: linkChildToParent
      tags: [parents]
      summary: Link a child account to parent
      parameters:
        - name: parentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkChildBody"
      responses:
        "201":
          description: Linked
  /students/{studentId}/enrollments:
    get:
      operationId: getStudentEnrollments
      tags: [students]
      summary: Get student's enrolled courses
      parameters:
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Enrollments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Enrollment"
components:
  schemas:
...
    RegisterBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
        role:
          type: string
          enum: [teacher, student, parent, admin]
        phone:
          type: string
        lang:
          type: string
          enum: [ar, en]
          default: ar
      required: [name, email, password, role]
    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required: [token, user]

    # ── USER ──
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        nameAr:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [teacher, student, parent, admin]
        avatarUrl:
          type: string
        phone:
          type: string
        lang:
          type: string
          enum: [ar, en]
        bio:
          type: string
        verified:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, role, createdAt]
...
    Course:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        titleAr:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        teacherId:
          type: integer
        teacherName:
          type: string
        teacherAvatarUrl:
          type: string
        subject:
          type: string
        gradeLevel:
          type: string
        price:
          type: number
        currency:
          type: string
          default: EGP
        coverImageUrl:
          type: string
        status:
          type: string
          enum: [draft, published, archived]
        enrollmentCount:
          type: integer
        rating:
          type: number
        lessonCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, title, teacherId, subject, status, createdAt]
...
    Enrollment:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        courseId:
          type: integer
        course:
          $ref: "#/components/schemas/Course"
        enrolledAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [active, completed, dropped]
        progress:
          type: number
      required: [id, studentId, courseId, enrolledAt, status]
...
    Assignment:
      type: object
      properties:
        id:
          type: integer
        courseId:
          type: integer
        title:
          type: string
        titleAr:
          type: string
        description:
          type: string
        dueDate:
          type: string
          format: date-time
        maxGrade:
          type: number
        submissionCount:
          type: integer
        gradedCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, courseId, title, dueDate, maxGrade, createdAt]
    CreateAssignmentBody:
      type: object
      properties:
        title:
          type: string
        titleAr:
          type: string
        description:
          type: string
        dueDate:
          type: string
          format: date-time
        maxGrade:
          type: number
      required: [title, dueDate, maxGrade]
    UpdateAssignmentBody:
      type: object
      properties:
        title:
          type: string
        titleAr:
          type: string
        description:
          type: string
        dueDate:
          type: string
          format: date-time
        maxGrade:
          type: number

    # ── SUBMISSION ──
    Submission:
      type: object
      properties:
        id:
          type: integer
        assignmentId:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        content:
          type: string
        fileUrl:
          type: string
        grade:
          type: number
        feedback:
          type: string
        status:
          type: string
          enum: [submitted, graded, returned]
        submittedAt:
          type: string
          format: date-time
        gradedAt:
          type: string
          format: date-time
      required: [id, assignmentId, studentId, content, status, submittedAt]
    CreateSubmissionBody:
      type: object
      properties:
        studentId:
          type: integer
        content:
          type: string
        fileUrl:
          type: string
      required: [studentId, content]
    GradeSubmissionBody:
      type: object
      properties:
        grade:
          type: number
        feedback:
          type: string
      required: [grade]

    # ── ANNOUNCEMENT ──
...
    DashboardSummary:
      type: object
      properties:
        totalStudents:
          type: integer
        totalCourses:
          type: integer
        totalLessons:
          type: integer
        totalRevenue:
          type: number
        pendingAssignments:
          type: integer
        unreadMessages:
          type: integer
        upcomingLessons:
          type: integer
        weeklyGrowth:
          type: number
        activeEnrollments:
          type: integer
      required: [totalStudents, totalCourses, totalLessons, totalRevenue, pendingAssignments, unreadMessages, upcomingLessons, weeklyGrowth, activeEnrollments]
    StudentProgress:
      type: object
      properties:
        courseId:
          type: integer
        courseName:
          type: string
        completedLessons:
          type: integer
        totalLessons:
          type: integer
        averageGrade:
          type: number
        progressPercent:
          type: number
        lastActivity:
          type: string
          format: date-time
      required: [courseId, courseName, completedLessons, totalLessons, progressPercent]
    RevenuePoint:
      type: object
      properties:
        month:
          type: string
        revenue:
          type: number
        enrollments:
          type: integer
      required: [month, revenue, enrollments]
    ActivityItem:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [enrollment, submission, grade, message, announcement, payment]
        actorName:
          type: string
        actorAvatarUrl:
          type: string
        description:
          type: string
        descriptionAr:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, type, actorName, description, createdAt]
    LinkChildBody:
      type: object
      properties:
        childId:
          type: integer
        linkCode:
          type: string
      required: [childId]